refactor: separate audio and server modules
